Removed MalcolmConnection
http://jira.diamond.ac.uk/browse/DAQ-301
