+++ org.eclipse.scanning.api/src/org/eclipse/scanning/api/event/scan/DeviceAction.java
@@ -31,8 +31,12 @@ public enum DeviceAction {
 	 * NOTE This does not make it run in a scan directly, it just marks it as being included
 	 * in new scans submitted to the server.
 	 */
-	ACTIVATE;
-
+	ACTIVATE,
+	
+	/**
+	 * Disable device, stopping all activity.
+	 */
+	DISABLE;
 	
 	public static DeviceAction as(TerminationPreference pref) {
 		switch(pref) {
