+++ org.eclipse.scanning.malcolm.core/src/org/eclipse/scanning/malcolm/core/MalcolmDevice.java
@@ -120,8 +120,6 @@ class MalcolmDevice<T> extends AbstractMalcolmDevice<T> {
 		
 		MalcolmMessage msg      = e.getBean();
 		DeviceState newState = MalcolmUtil.getState(msg, false);
-		
-		System.out.println("######MTMT####### SCAN EVENT RECEIVED: " + newState); // TODO REMOVE
 
 		ScanBean bean = getBean();
 		bean.setDeviceName(getName());
@@ -145,8 +143,6 @@ class MalcolmDevice<T> extends AbstractMalcolmDevice<T> {
 		MalcolmMessage msg = e.getBean();
 		
 		DeviceState newState = MalcolmUtil.getState(msg, false);
-		
-		System.out.println("######MTMT####### STATE CHANGED RECEIVED: " + newState); // TODO REMOVE
 		
 		// Send scan state changed
 		ScanBean bean = getBean();
@@ -204,8 +200,10 @@ class MalcolmDevice<T> extends AbstractMalcolmDevice<T> {
 	@Override
 	public void configure(T model) throws MalcolmDeviceException {
 		final MalcolmMessage msg   = connectionDelegate.createCallMessage("configure", model);
-		service.send(this, msg);
-		// TODO handle return from configure. Check for error?
+		MalcolmMessage reply = service.send(this, msg);
+        if (reply.getType() == Type.ERROR) {
+        	throw new MalcolmDeviceException(reply.getMessage());
+        }
 		setModel(model);
 	}
 
@@ -216,6 +214,11 @@ class MalcolmDevice<T> extends AbstractMalcolmDevice<T> {
 
 	@Override
 	public void abort() throws MalcolmDeviceException {
+		connectionDelegate.call(Thread.currentThread().getStackTrace());
+	}
+
+	@Override
+	public void disable() throws MalcolmDeviceException {
 		connectionDelegate.call(Thread.currentThread().getStackTrace());
 	}
