+++ org.eclipse.scanning.malcolm.core/src/org/eclipse/scanning/malcolm/core/MalcolmService.java
@@ -1,30 +1,104 @@
 package org.eclipse.scanning.malcolm.core;
 
-import java.net.URI;
-import java.net.URISyntaxException;
+import java.util.Collection;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
-import org.eclipse.scanning.api.malcolm.IMalcolmConnection;
+import org.eclipse.scanning.api.event.core.IPublisher;
+import org.eclipse.scanning.api.event.scan.ScanBean;
+import org.eclipse.scanning.api.malcolm.IMalcolmDevice;
 import org.eclipse.scanning.api.malcolm.IMalcolmService;
 import org.eclipse.scanning.api.malcolm.MalcolmDeviceException;
 import org.eclipse.scanning.api.malcolm.connector.IMalcolmConnectorService;
 import org.eclipse.scanning.api.malcolm.message.MalcolmMessage;
 
 public class MalcolmService implements IMalcolmService {
-		
+
+	private IMalcolmConnectorService<MalcolmMessage> connector;	
+	private Map<String, IMalcolmDevice<?>>        devices;
+	
 	/**
 	 * Used by OSGi to make the service.
 	 */
-	public MalcolmService() {
+	public MalcolmService() throws MalcolmDeviceException {
+		this(createConnectorService());
 	}
 
-	@Override
-	public IMalcolmConnection createConnection(URI malcolmUri) throws URISyntaxException, MalcolmDeviceException {
-		return new MalcolmConnection(malcolmUri);
+	/**
+	 * 
+	 * @param connectorService - null in the OSGi services scenario
+	 * @param malcolmUri - 
+	 * @throws MalcolmDeviceException
+	 */
+	public MalcolmService(IMalcolmConnectorService<MalcolmMessage> connector) throws MalcolmDeviceException {
+		
+		this.connector = connector; // Usually null unless we are in testing mode.
+		this.devices   = new ConcurrentHashMap<String, IMalcolmDevice<?>>(4);
+		
+		// At this point we get OSGi to get the connector service for us.
+		try {
+			if (connector==null) throw new MalcolmDeviceException("Cannot find an implementation of IConnectorService!");
+			connector.connect(null);
+		} catch (Exception e) {
+			throw new MalcolmDeviceException(null, "Cannot get connection services!", e);
+		}
 	}
 
+	/**
+	 * Gets the service from OSGi
+	 * @return The connector service
+	 * @throws MalcolmDeviceException
+	 */
+	private static IMalcolmConnectorService<MalcolmMessage> createConnectorService() throws MalcolmDeviceException {
+		
+		try {
+			Collection<IMalcolmConnectorService<MalcolmMessage>> services = MalcolmActivator.getConnectionServices();
+			if (services.size()>1) throw new Exception("We have more than one connector service and not information to choose between them!"); // TODO
+	
+			return services.iterator().next();
+		} catch (Exception ne) {
+			throw new MalcolmDeviceException("Cannot find a service implementing IConnectorService!");
+		}
+	}
+	
+	/**
+	 * Get a device by name. At the point where the device is retrieved the
+	 * caller may know the type of device and use a generic to declare its model.
+	 * 
+	 * @param name
+	 * @return
+	 * @throws MalcolmDeviceException
+	 */
+	public <T> IMalcolmDevice<T> getDevice(String name) throws MalcolmDeviceException {
+        return getDevice(name, null);
+	}
+	
+	/**
+	 * Get a device by name. At the point where the device is retrieved the
+	 * caller may know the type of device and use a generic to declare its model.
+	 * 
+	 * @param name
+	 * @param publisher
+	 * @return
+	 * @throws MalcolmDeviceException
+	 */
+	@SuppressWarnings("unchecked")
+	public <T> IMalcolmDevice<T> getDevice(String name, IPublisher<ScanBean> publisher) throws MalcolmDeviceException {
+		
+		if (devices.containsKey(name)) return (IMalcolmDevice<T>)devices.get(name);
+		
+		IMalcolmDevice<T> device = new MalcolmDevice<T>(name, connector, publisher); // Might throw exception
+		devices.put(name, device);
+		return device;
+	}
 
+	/**
+	 * Disposes the service
+	 * @throws MalcolmDeviceException
+	 */
 	@Override
-	public IMalcolmConnection createConnection(URI malcolmUri, IMalcolmConnectorService<MalcolmMessage> connectorService) throws URISyntaxException, MalcolmDeviceException {
-		return new MalcolmConnection(malcolmUri, connectorService);
+	public void dispose() throws MalcolmDeviceException {
+		devices.clear();
+		connector.disconnect();
 	}
 }
