+++ org.eclipse.scanning.sequencer/src/org/eclipse/scanning/sequencer/AcquisitionDevice.java
@@ -384,6 +384,11 @@ final class AcquisitionDevice extends AbstractRunnableDevice<ScanModel> {
 			lock.unlock();
 		}
 	}
+	
+	@Override
+	public void disable() throws ScanningException {
+		// TODO Matt Gerring to implement this: call abort on all children
+	}
 
 	@Override
 	public void pause() throws ScanningException {
