+++ org.eclipse.scanning.test/src/org/eclipse/scanning/test/malcolm/mock/MockMultipleClientTest.java
@@ -1,6 +1,5 @@
 package org.eclipse.scanning.test.malcolm.mock;
 
-import org.eclipse.scanning.api.malcolm.IMalcolmConnection;
 import org.eclipse.scanning.api.malcolm.IMalcolmDevice;
 import org.eclipse.scanning.connector.epics.EpicsV4ConnectorService;
 import org.eclipse.scanning.test.malcolm.AbstractMultipleClientMalcolmTest;
@@ -16,23 +15,20 @@ public class MockMultipleClientTest extends AbstractMultipleClientMalcolmTest {
 	@Before
 	public void create() throws Exception {
 		this.connectorService = new EpicsV4ConnectorService();
-		this.service      = new MockedMalcolmService();
-		this.connection   = service.createConnection(PAUSABLE);
-		this.device       =  connection.getDevice("zebra");
+		this.service      = new MockedMalcolmService(true);
+		this.device       =  service.getDevice("zebra");
 	}
 
 	@Override
 	@After
 	public void dispose() throws Exception {
 		if (device!=null)     device.dispose();
-		if (connection!=null) connection.dispose();
 		((MockedMalcolmService)service).dispose();
 	}
 	
 	@Override
 	protected IMalcolmDevice createAdditionalConnection() throws Exception {
-		IMalcolmConnection aconnection   = service.createConnection(PAUSABLE);
-		return aconnection.getDevice("zebra");
+		return service.getDevice("zebra");
 	}
 
 }
